requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1718029820055/work
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
html5lib @ file:///Users/ktietz/demo/mc3/conda-bld/html5lib_1629144453894/work
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml @ file:///croot/lxml_1737039601731/work
MarkupSafe==3.0.2
multidict==6.4.4
openai==0.28.0
playwright==1.52.0
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
python-dotenv==1.1.0
requests==2.32.3
scraper==0.1.0
six @ file:///croot/six_1744271502820/work
sniffio==1.3.1
soupsieve @ file:///croot/soupsieve_1696347547217/work
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
yarl==1.20.0
